@@ -20,9 +20,10 @@
     "dependencies": {
         "@emailjs/browser": "^4.3.1",
         "@mapbox/rehype-prism": "^0.9.0",
-        "@tabler/icons-svelte": "^3.2.0",
+        "@tabler/icons-svelte": "^3.12.0",
         "@vercel/analytics": "^1.2.2",
         "@vercel/speed-insights": "^1.0.10",
+        "clsx": "^2.1.1",
         "front-matter": "^4.0.2",
         "glob": "^10.3.12",
         "rehype": "^13.0.1",
@@ -33,6 +34,7 @@
         "remark-parse": "^11.0.0",
         "remark-rehype": "^11.1.0",
         "shiki": "^1.5.0",
+        "tailwind-merge": "^2.5.2",
         "typesafe-i18n": "^5.26.2",
         "unified": "^11.0.4"
     },
@@ -59,7 +61,7 @@
         "postcss": "^8.4.35",
         "prettier": "^3.2.5",
         "prettier-plugin-svelte": "^3.2.2",
-        "svelte": "^4.2.12",
+        "svelte": "^5.0.0-next.238",
         "svelte-check": "^3.6.6",
         "tailwindcss": "^3.4.1",
         "tslib": "^2.6.2",
