@@ -1,13 +1,17 @@
 <script lang="ts">
-    export let type: 'text' | 'email' = 'text';
-    export let value: string;
-    export let placeholder: string;
+    interface Props {
+        placeholder: string;
+        type?: 'text' | 'email';
+        value: string;
+    }
+
+    let { placeholder, type = 'text', value = $bindable() }: Props = $props();
 </script>
 
 <input
     {type}
     {value}
-    on:change={event => (value = event.currentTarget.value)}
     {placeholder}
+    onchange={(event) => (value = event.currentTarget.value)}
     class="relative w-full h-[50px] rounded bg-gray-100 dark:bg-zinc-700 px-5 text-[14px] resize-none"
 />
