@@ -1,25 +1,28 @@
 <script lang="ts">
-    import SuccessIcon from '@tabler/icons-svelte/IconCheck.svelte';
-    import ErrorIcon from '@tabler/icons-svelte/IconExclamationMark.svelte';
-    import InfoIcon from '@tabler/icons-svelte/IconInfoCircle.svelte';
-    import CloseIcon from '@tabler/icons-svelte/IconX.svelte';
-    import { createEventDispatcher } from 'svelte';
     import { fade } from 'svelte/transition';
+    import CloseIcon from '@tabler/icons-svelte/icons/x';
+    import { BG_COLOR_MAP, ICONS_MAP } from './defaults';
+    import type { Snippet } from 'svelte';
+    import type { Toast } from '$stores/toast/index.svelte';
 
-    const dispatch = createEventDispatcher<{ dismiss: null }>();
+    interface Props {
+        type: Toast["type"],
+        ondismiss: () => void,
+        children: Snippet
+    }
 
-    export let type: 'success' | 'error' | 'info';
+    let { children, ondismiss, type }: Props = $props();
 
-    $: icon = type === 'success' ? SuccessIcon : type === 'error' ? ErrorIcon : InfoIcon;
-    $: bgColor = type === 'success' ? 'bg-green-600' : type === 'error' ? 'bg-red-500' : 'bg-blue-500';
+    let Icon = $derived(ICONS_MAP[type]);
+    let bgColor = $derived(BG_COLOR_MAP[type]);
 </script>
 
 <article class="relative w-full flex items-center gap-4 rounded px-4 py-2 text-neutral-50 {bgColor}" role="alert" transition:fade>
-    <svelte:component this={icon} class="size-6" />
+    <Icon class="min-h-6 min-w-6" />
     <p class="text-sm flex-grow">
-        <slot />
+        {@render children()}
     </p>
-    <button class="border-none bg-transparent text-neutral-50" on:click={() => dispatch('dismiss')}>
+    <button class="border-none bg-transparent text-neutral-50" onclick={ondismiss}>
         <CloseIcon class="size-6" />
     </button>
 </article>
