@@ -9,8 +9,8 @@
     import { goto, invalidateAll } from '$app/navigation';
     import { page } from '$app/stores';
 
-    let showList = false;
-    let currentLanguageNode: HTMLButtonElement;
+    let showList = $state(false);
+    let currentLanguageNode: HTMLButtonElement | undefined = $state<HTMLButtonElement>();
 
     function isLocale(locale: string): locale is Locales {
         return locales.includes(locale as Locales);
@@ -30,15 +30,18 @@
         await invalidateAll();
     };
 
-    // update `lang` attribute
-    $: browser && document.querySelector('html')!.setAttribute('lang', $locale);
+    $effect(() => {
+        // update `lang` attribute
+        browser && document.querySelector('html')!.setAttribute('lang', $locale);
+    });
 
-    // update locale when page store changes
-    $: if (browser) {
-        const lang = $page.params.lang as Locales;
-        switchLocale(lang, false);
-        history.replaceState({ ...history.state, locale: lang }, '', replaceLocaleInUrl($page.url, lang));
-    }
+    $effect(() => {
+        if (browser) {
+            const lang = $page.params.lang as Locales;
+            switchLocale(lang, false);
+            history.replaceState({ ...history.state, locale: lang }, '', replaceLocaleInUrl($page.url, lang));
+        }
+    });
 
     function changeLang(newLocale: string) {
         showList = false;
@@ -49,23 +52,23 @@
 <div class="relative">
     <button
         bind:this={currentLanguageNode}
-        on:click={() => (showList = !showList)}
+        onclick={() => (showList = !showList)}
         class="bg-gray-700 px-4 py-2 rounded {showList ? 'rounded-b-none' : ''}"
     >
         {$locale === 'fr' ? $LL.languages.french() : $LL.languages.english()}
     </button>
     {#if showList}
         <div
-            on:emit={() => (showList = false)}
+            onemit={() => (showList = false)}
             use:clickOutside={{ avoid: [currentLanguageNode] }}
             class="absolute w-full top-full left-0 flex flex-col z-20"
         >
             {#if $locale === 'en'}
-                <button on:click={() => changeLang('fr')} class="relative w-full bg-gray-700 px-4 py-2 rounded {showList ? 'rounded-t-none' : ''}">
+                <button onclick={() => changeLang('fr')} class="relative w-full bg-gray-700 px-4 py-2 rounded {showList ? 'rounded-t-none' : ''}">
                     {$LL.languages.french()}
                 </button>
             {:else}
-                <button on:click={() => changeLang('en')} class="relative w-full bg-gray-700 px-4 py-2 rounded {showList ? 'rounded-t-none' : ''}">
+                <button onclick={() => changeLang('en')} class="relative w-full bg-gray-700 px-4 py-2 rounded {showList ? 'rounded-t-none' : ''}">
                     {$LL.languages.english()}
                 </button>
             {/if}
