@@ -1,11 +1,10 @@
 <script lang="ts">
     import '../app.css';
     import { onMount } from 'svelte';
-    import { initTheme } from '$lib/stores/theme';
+    import { initTheme } from '$stores/theme/index.svelte';
     import { setLocale } from '../i18n/i18n-svelte';
-    import type { LayoutData } from './$types';
 
-    export let data: LayoutData;
+    let { children, data } = $props();
 
     setLocale(data.locale);
     
@@ -14,4 +13,4 @@
     });
 </script>
 
-<slot />
+{@render children()}
